refactor instructor and person classes
